+++ plugin-locationcomponent/src/main/java/com/mapbox/maps/plugin/locationcomponent/LayerSourceProvider.kt
@@ -1,12 +1,14 @@
 package com.mapbox.maps.plugin.locationcomponent
 
 import android.content.Context
+import com.mapbox.bindgen.Value
 import com.mapbox.maps.MapboxExperimental
 import com.mapbox.maps.plugin.LocationPuck2D
 import com.mapbox.maps.plugin.LocationPuck3D
 import com.mapbox.maps.plugin.locationcomponent.LocationComponentConstants.LOCATION_INDICATOR_LAYER
 import com.mapbox.maps.plugin.locationcomponent.LocationComponentConstants.MODEL_LAYER
 import com.mapbox.maps.plugin.locationcomponent.LocationComponentConstants.MODEL_SOURCE
+import com.mapbox.maps.plugin.locationcomponent.utils.take
 import java.lang.ref.WeakReference
 
 @OptIn(MapboxExperimental::class)
@@ -34,6 +36,8 @@ internal object LayerSourceProvider {
       locationModelLayerOptions.modelReceiveShadows,
       locationModelLayerOptions.modelOpacity.toDouble(),
       locationModelLayerOptions.modelScaleMode,
+      locationModelLayerOptions.modelEmissiveStrength.toDouble(),
+      locationModelLayerOptions.modelEmissiveStrengthExpression?.let { Value.fromJson(it).take() }
     )
 
   fun getLocationIndicatorLayer() = LocationIndicatorLayerWrapper(LOCATION_INDICATOR_LAYER)